requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
idna==3.7
mysql==0.0.3
mysqlclient==2.2.4
pipenv==2023.12.1
platformdirs==4.2.1
psycopg2==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.5.1
soupsieve==2.6
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.1
